+++ app/build.gradle
@@ -66,14 +66,13 @@ dependencies {
     implementation 'com.google.android.material:material:1.0.0'
     implementation "com.google.android.gms:play-services-location:15.0.1"
     implementation 'com.getbase:floatingactionbutton:1.9.1'
+    implementation 'com.googlecode.json-simple:json-simple:1.1.1'
     testImplementation 'junit:junit:4.13'
-    testImplementation 'org.mockito:mockito-core:3.3.0'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.uiautomator:uiautomator:2.2.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     androidTestImplementation 'androidx.test:rules:1.2.0'
-    implementation 'com.googlecode.json-simple:json-simple:1.1.1'
-    androidTestImplementation"com.android.support.test.espresso:espresso-intents:3.2.0"
+    androidTestImplementation "androidx.test.espresso:espresso-intents:3.2.0"
     androidTestImplementation 'androidx.test:rules:1.1.1'
     androidTestImplementation 'androidx.test:runner:1.1.1'
 }
