+++ app/src/main/assets/buildings.json
